[feat] #55 SettingMenuType 따로 추출
